pyproject.toml
[tool.poetry]
name = "news_parser"
version = "0.1.0"
description = ""
authors = ["Maxim Vasilev <vsmaxim8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
newspaper3k = "^0.2.8"
lxml = "^4.5.0"
SQLAlchemy = "^1.3.13"
psycopg2-binary = "^2.8.4"
vk_api = "^11.7.0"
pyyaml = "^5.3"
deeppavlov = "^0.8.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
